Convert all kind enums from u32 to i32
It appears as if bindgen chooses i32 always for Windows, but varies for Mac and
Linux. Relevant links:

- https://users.rust-lang.org/t/bindgen-and-enum-types/34370
- https://github.com/rust-lang/rust-bindgen/issues/1361

However, this behavior isn't in line with the C standard, which states:

"The identifiers in an enumerator list are declared as constants that have type
int and may appear wherever such are permitted."

To remedy this for all platforms, enums are now default type i32.

+++ src/base.rs
@@ -131,7 +131,7 @@ impl Rs2Intrinsics {
     /// Distortion model and coefficients of the image
     pub fn distortion(&self) -> Rs2Distortion {
         Rs2Distortion {
-            model: Rs2DistortionModel::from_i32(self.0.model).unwrap(),
+            model: Rs2DistortionModel::from_i32(self.0.model as i32).unwrap(),
             coeffs: self.0.coeffs,
         }
     }
